Move game bootstrap inputs outward
